+++ dzVents/runtime/Domoticz.lua
@@ -146,7 +146,6 @@ local function Domoticz(settings)
 		['BASETYPE_TIMER'] = 'timer',
 		['BASETYPE_HTTP_RESPONSE'] = 'httpResponse',
 
-
 		utils = {
 			_ = _,
 
@@ -189,7 +188,7 @@ local function Domoticz(settings)
 			hsbToRGB = function(h, s, b)
 				return utils.hsbToRGB(h,s,b)
 			end,
-			
+
 			dumpTable = function(t, level)
 				return utils.dumpTable(t, level)
 			end,
@@ -201,6 +200,47 @@ local function Domoticz(settings)
 			inTable = function(t, searchItem)
 				return utils.inTable(t, searchItem)
 			end,
+
+			rightPad = function(text, length, char)
+				return utils.rightPad(text, length, char)
+			end,
+
+			leftPad = function(text, length, char)
+				return utils.leftPad(text, length, char)
+			end,
+
+			centerPad = function(text, length, char)
+				return utils.centerPad(text, length, char)
+			end,
+
+			leadingZeros = function(num, length)
+				return utils.leadingZeros(num, length)
+			end,
+
+			numDecimals = function(num, int, dec)
+				return utils.numDecimals(num, int, dec)
+			end,
+
+			deviceExists = function (parm)
+				return utils.deviceExists(parm)
+			end,
+
+			sceneExists = function (parm)
+				return utils.sceneExists(parm)
+			end,
+
+			groupExists = function (parm)
+				return utils.groupExists(parm)
+			end,
+
+			variableExists = function (parm)
+				return utils.variableExists(parm)
+			end,
+
+			cameraExists = function (parm)
+				return utils.cameraExists(parm)
+			end,
+
 		}
 	}
 
@@ -236,11 +276,11 @@ local function Domoticz(settings)
 		end
 
 		--[[
-        if _subSystem:find('gcm') then
+		if _subSystem:find('gcm') then
 			utils.log('Notification subsystem Google Cloud Messaging (gcm) has been deprecated by Google. Please consider switching to Firebase', utils.LOG_ERROR)
 		end
-        ]] --
-        
+		]] --
+
 		local data = subject
 				.. '#' .. message
 				.. '#' .. tostring(priority)
@@ -502,7 +542,6 @@ local function Domoticz(settings)
 					_item = item
 				end
 
-
 				if (_item and type(_item) ~= 'function' and ((initial == true and type(i) == 'number') or (initial == false and type(i) ~= number))) then
 					res = func(_item)
 					if (res == false) then -- abort
@@ -526,7 +565,6 @@ local function Domoticz(settings)
 				else
 					_item = item
 				end
-
 
 				if (_item and type(_item) ~= 'function' and ((initial == true and type(i) == 'number') or (initial == false and type(i) ~= number))) then
 					ret = func(_item)